+++ LittleWigs_BattleForAzeroth/HorrificVisionOfStormwind/Trash.lua
@@ -50,6 +50,7 @@ mod:RegisterEnableMob(
 --
 
 local portalsClosed = 0
+local activeBuffs = {}
 
 --------------------------------------------------------------------------------
 -- Localization
@@ -64,6 +65,10 @@ if L then
 	L.madnesses = "Madnesses"
 	L.potions = "Potions"
 	L.buffs = "Buffs"
+	L.slowed = "Slowed"
+	L.sluggish_potion_effect = "Heal 2% every 5 sec"
+	L.sickening_potion_effect = "5% damage reduction"
+	L.spicy_potion_effect = "Breathe fire"
 
 	L.crawling_corruption = "Crawling Corruption"
 	L.enthralled_footman = "Enthralled Footman"
@@ -118,7 +123,11 @@ function mod:GetOptions()
 		-- Potions
 		315814, -- Fermented Mixture
 		315807, -- Noxious Mixture
+		315845, -- Sluggish Potion
+		315849, -- Sickening Potion
+		315817, -- Spicy Potion
 		-- Buffs
+		313698, -- Gift of the Titans
 		312456, -- Elite Extermination
 		314203, -- Requited Bulwark
 		312355, -- Bear Spirit
@@ -188,7 +197,7 @@ function mod:GetOptions()
 		["altpower"] = "general",
 		[311390] = L.madnesses,
 		[315814] = L.potions,
-		[312456] = L.buffs,
+		[313698] = L.buffs,
 		[296510] = L.crawling_corruption,
 		[298584] = L.enthralled_footman,
 		[308375] = L.fallen_voidspeaker,
@@ -239,6 +248,8 @@ function mod:OnBossEnable()
 	-- Potions
 	self:Log("SPELL_ENERGIZE", "FermentedMixture", 315814)
 	self:Log("SPELL_ENERGIZE", "NoxiousMixture", 315807)
+	self:InitBuffs() -- reload protection
+	self:RegisterUnitEvent("UNIT_AURA", nil, "player")
 
 	-- Buffs
 	self:Log("SPELL_ENERGIZE", "EliteExtermination", 312456)
@@ -399,6 +410,7 @@ end
 
 function mod:OnBossDisable()
 	portalsClosed = 0
+	activeBuffs = {}
 end
 
 --------------------------------------------------------------------------------
@@ -516,6 +528,111 @@ function mod:NoxiousMixture(args)
 		local sanityLost = args.extraSpellId -- will be a negative number representing Sanity lost
 		self:Message(args.spellId, "yellow", CL.other:format(args.spellName, L.sanity_change:format(sanityLost)))
 		self:PlaySound(args.spellId, "warning")
+	end
+end
+
+do
+	local buffIds = {
+		[315845] = true, -- Sluggish Potion
+		[315849] = true, -- Sickening Potion
+		[315817] = true, -- Spicy Potion
+		[313698] = true, -- Gift of the Titans
+	}
+	local numBuffs = #buffIds
+
+	function mod:InitBuffs() -- reload protection
+		for spellId in next, buffIds do
+			local auraTbl = self:GetPlayerAura(spellId)
+			if auraTbl then
+				activeBuffs[auraTbl.auraInstanceID] = {auraTbl.expirationTime, auraTbl.spellId}
+				if spellId == 315845 then -- Sluggish Potion
+					self:Bar(spellId, auraTbl.expirationTime - GetTime(), L.sluggish_potion_effect)
+				elseif spellId == 315849 then -- Sickening Potion
+					self:Bar(spellId, auraTbl.expirationTime - GetTime(), L.sickening_potion_effect)
+				elseif spellId == 315817 then -- Spicy Potion
+					self:Bar(spellId, auraTbl.expirationTime - GetTime(), L.spicy_potion_effect)
+				elseif spellId == 313698 then -- Gift of the Titans
+					self:Bar(spellId, auraTbl.expirationTime - GetTime())
+				end
+			end
+		end
+	end
+
+	function mod:UNIT_AURA(_, _, updateInfo)
+		if not updateInfo or updateInfo.isFullUpdate then
+			self:InitBuffs()
+		else
+			if updateInfo.addedAuras then
+				for i = 1, #updateInfo.addedAuras do
+					local auraTbl = updateInfo.addedAuras[i]
+					local spellId = auraTbl.spellId
+
+					if buffIds[spellId] then
+						buffIds[auraTbl.auraInstanceID] = {auraTbl.expirationTime, spellId}
+						if spellId == 315845 then -- Sluggish Potion
+							self:Message(spellId, "green", CL.other:format(CL.you:format(auraTbl.name), L.sluggish_potion_effect))
+							self:Bar(spellId, auraTbl.expirationTime - GetTime(), L.sluggish_potion_effect)
+							self:PlaySound(spellId, "info")
+						elseif spellId == 315849 then -- Sickening Potion
+							self:Message(spellId, "green", CL.other:format(CL.you:format(auraTbl.name), L.sickening_potion_effect))
+							self:Bar(spellId, auraTbl.expirationTime - GetTime(), L.sickening_potion_effect)
+							self:PlaySound(spellId, "info")
+						elseif spellId == 315817 then -- Spicy Potion
+							self:Message(spellId, "green", CL.other:format(CL.you:format(auraTbl.name), L.spicy_potion_effect))
+							self:Bar(spellId, auraTbl.expirationTime - GetTime(), L.spicy_potion_effect)
+							self:PlaySound(spellId, "info")
+						elseif spellId == 313698 then -- Gift of the Titans
+							self:Message(spellId, "green", CL.you:format(auraTbl.name))
+							self:Bar(spellId, auraTbl.expirationTime - GetTime())
+							self:PlaySound(spellId, "long")
+						end
+					end
+				end
+			end
+			if updateInfo.removedAuraInstanceIDs then
+				for i = 1, #updateInfo.removedAuraInstanceIDs do
+					local hadBuff = buffIds[updateInfo.removedAuraInstanceIDs[i]]
+					if hadBuff then
+						local spellId = hadBuff[2]
+						buffIds[updateInfo.removedAuraInstanceIDs[i]] = nil
+						if spellId == 315845 then -- Sluggish Potion
+							self:Message(spellId, "blue", CL.other:format(CL.removed:format(self:SpellName(spellId)), L.slowed))
+							self:StopBar(L.sluggish_potion_effect)
+							self:PlaySound(spellId, "warning")
+						elseif spellId == 315849 then -- Sickening Potion
+							self:Message(spellId, "blue", CL.other:format(CL.removed:format(self:SpellName(spellId)), self:SpellName(315850))) -- Vomit
+							self:StopBar(L.sickening_potion_effect)
+							self:PlaySound(spellId, "warning")
+						elseif spellId == 315817 then -- Spicy Potion
+							self:Message(spellId, "blue", CL.other:format(CL.removed:format(self:SpellName(spellId)), self:SpellName(315818))) -- Burning
+							self:StopBar(L.spicy_potion_effect)
+							self:PlaySound(spellId, "warning")
+						end
+					end
+				end
+			end
+			if updateInfo.updatedAuraInstanceIDs then
+				for i = 1, #updateInfo.updatedAuraInstanceIDs do
+					local hadBuff = buffIds[updateInfo.updatedAuraInstanceIDs[i]]
+					if hadBuff then
+						local spellId = hadBuff[2]
+						local auraTbl = self:GetPlayerAura(spellId)
+						if hadBuff[1] ~= auraTbl.expirationTime then
+							buffIds[updateInfo.updatedAuraInstanceIDs[i]][1] = auraTbl.expirationTime
+							if spellId == 315845 then -- Sluggish Potion
+								self:Bar(spellId, auraTbl.expirationTime - GetTime(), L.sluggish_potion_effect)
+							elseif spellId == 315849 then -- Sickening Potion
+								self:Bar(spellId, auraTbl.expirationTime - GetTime(), L.sickening_potion_effect)
+							elseif spellId == 315817 then -- Spicy Potion
+								self:Bar(spellId, auraTbl.expirationTime - GetTime(), L.spicy_potion_effect)
+							elseif spellId == 313698 then -- Gift of the Titans
+								self:Bar(spellId, auraTbl.expirationTime - GetTime())
+							end
+						end
+					end
+				end
+			end
+		end
 	end
 end