+++ SpellActivationOverlay/classes/mage.lua
@@ -14,11 +14,13 @@ local clearcastingVariants; -- Lazy init in lazyCreateClearcastingVariants()
 
 local arcaneMissiles = 5143;
 local fireBlast = 2136;
+local frostfireBolt = 44614;
+local infernoBlast = 108853; -- Replaces Fire Blast in Mists of Pandaria
 local pyroblast = 11366; -- Pyroblast, the base Pyro spell
 local pyroblastBang = 92315; -- Pyroblast!, a specific spell for instant Pyro introduced in Cataclysm
 
 local hotStreakSpellID = 48108;
-local heatingUpSpellID = 48107; -- Does not exist in WoW Classic yet
+local heatingUpSpellID = 48107; -- Does not exist in WoW Classic until MoP
 local hotStreakHeatingUpSpellID = hotStreakSpellID+heatingUpSpellID; -- Made up entirely, does not even exist in Retail
 local improvedHotStreakSpellID = 44446; -- Cataclysm talent that trigger Heating Up
 local hotStreakSoDSpellID = 400625;
@@ -34,7 +36,7 @@ HotStreakHandler.init = function(self, talentName)
     local fire_blast = { 2136, 2137, 2138, 8412, 8413, 10197, 10199, 27078, 27079, 42872, 42873 }
     local fire_blast_sod = { 400618, 400619, 400616, 400620, 400621, 400622, 400623 } -- Improved by Overheat rune
     local fireball = { 133, 143, 145, 3140, 8400, 8401, 8402, 10148, 10149, 10150, 10151, 25306, 27070, 38692, 42832, 42833 }
-    local frostfire_bolt = { 44614, 47610 }
+    local frostfire_bolt = { frostfireBolt, 47610 }
     local frostfire_bolt_sod = { 401502 }
     -- local living_bomb = { 44457, 55359, 55360 } this is the DOT effect, which we do NOT want
     local living_bomb = { 44461, 55361, 55362 }
@@ -469,7 +471,10 @@ local function customLogin(self, ...)
     -- Must initialize class on PLAYER_LOGIN instead of registerClass
     -- Because we need the talent tree, which is not always available right off the bat
     local hotStreakSpellName;
-    if SAO.IsSoD() then
+    if SAO.IsProject(SAO.MOP_AND_ONWARD) then
+        -- Invalidate hotStreakSpellName to avoid using HotStreakHandler, which is deprecated from Mists of Pandaria
+        hotStreakSpellName = nil;
+    elseif SAO.IsSoD() then
         hotStreakSpellName = GetSpellInfo(hotStreakSoDSpellID);
     elseif SAO.IsCata() then
         hotStreakSpellName = GetSpellInfo(improvedHotStreakSpellID);
@@ -515,8 +520,8 @@ local function lazyCreateClearcastingVariants(self)
         return;
     end
 
-    if not self.IsProject(SAO.ERA + SAO.TBC + SAO.WRATH + SAO.CATA) then
-        -- Clearcasting exists up until Cataclysm
+    if self.IsProject(SAO.MOP_AND_ONWARD) then
+        -- Clearcasting stopped existing since Mists of Pandaria
         return;
     end
 
@@ -558,27 +563,104 @@ local function useImpact()
 end
 
 local function useArcaneMissiles()
-    local arcaneMissilesBuff = 79683; -- Cataclysm requires a buff before casting Arcane Missiles
+    local arcaneMissilesBuff = 79683; -- Cataclysm and MoP require a buff before casting Arcane Missiles
 
+    if SAO.IsCata() then
+        SAO:CreateEffect(
+            "arcane_missiles",
+            SAO.CATA,
+            arcaneMissilesBuff,
+            "aura",
+            {
+                overlay = { texture = "arcane_missiles", position = "Left + Right (Flipped)", scale = 0.6 }, -- Smaller, to avoid overlap with Arcane Potency
+                button = arcaneMissiles,
+                handler = {
+                    -- Force refresh on a regular basis, because the game client does not send the correct SPELL_AURA_REFRESH events
+                    -- We can call refresh() even without an active display; the bucket will send refresh to displays if and only if there is an active one
+                    onRepeat = function(bucket) bucket:refresh(); end,
+                }
+            }
+        );
+    elseif SAO.IsMoP() then
+        local hash0Stacks = SAO:HashNameFromStacks(0);
+        local hash2Stacks = SAO:HashNameFromStacks(2);
+        SAO:CreateEffect(
+            "arcane_missiles",
+            SAO.MOP,
+            arcaneMissilesBuff,
+            "aura",
+            {
+                overlays = {
+                    { stacks = 1, texture = "arcane_missiles", position = "Left", option = false },
+                    { stacks = 2, texture = "arcane_missiles", position = "Left + Right (Flipped)", option = { setupHash = hash0Stacks, testHash = hash2Stacks } },
+                },
+                button = arcaneMissiles,
+            }
+        );
+    end
+end
+
+local function useFingersOfFrost()
+    local fingersOfFrostBuff = 44544;
+    local hash0Stacks = SAO:HashNameFromStacks(0);
+    local hash2Stacks = SAO:HashNameFromStacks(2);
     SAO:CreateEffect(
-        "arcane_missiles",
-        SAO.CATA,
-        arcaneMissilesBuff,
+        "fingers_of_frost",
+        SAO.MOP,
+        fingersOfFrostBuff,
         "aura",
         {
-            overlay = { texture = "arcane_missiles", position = "Left + Right (Flipped)", scale = 0.6 }, -- Smaller, to avoid overlap with Arcane Potency
-            button = arcaneMissiles,
-            handler = {
-                -- Force refresh on a regular basis, because the game client does not send the correct SPELL_AURA_REFRESH events
-                -- We can call refresh() even without an active display; the bucket will send refresh to displays if and only if there is an active one
-                onRepeat = function(bucket) bucket:refresh(); end,
+            overlays = { -- Slightly bigger to avoid overlap with Arcane Missiles, and slightly dimmer to compensate
+                { stacks = 1, texture = "frozen_fingers", position = "Left",                   scale = 1.1, color = { 222, 222, 222 }, option = false },
+                { stacks = 2, texture = "frozen_fingers", position = "Left + Right (Flipped)", scale = 1.1, color = { 222, 222, 222 }, option = { setupHash = hash0Stacks, testHash = hash2Stacks } },
+            },
+            buttons = {
+                FrozenHandler.ice_lance[1],
+                FrozenHandler.deep_freeze[1],
             },
         }
     );
 end
 
-local function registerClass(self)
-    -- Fire Procs
+local function useBrainFreeze()
+    local brainFreezeBuff = 57761;
+    SAO:CreateEffect(
+        "brain_freeze",
+        SAO.MOP,
+        brainFreezeBuff,
+        "aura",
+        {
+            overlay = { texture = "brain_freeze", position = "Top" },
+            button = frostfireBolt,
+        }
+    );
+end
+
+local function useHeatingUpAndHotStreak()
+    SAO:CreateEffect(
+        "heating_up",
+        SAO.MOP,
+        heatingUpSpellID,
+        "aura",
+        {
+            overlay = { texture = "hot_streak", position = "Left + Right (Flipped)", scale = 0.5 },
+            button = infernoBlast,
+        }
+    );
+
+    SAO:CreateEffect(
+        "hot_streak",
+        SAO.MOP,
+        hotStreakSpellID,
+        "aura",
+        {
+            overlay = { texture = "hot_streak", position = "Left + Right (Flipped)" },
+            button = pyroblast,
+        }
+    );
+end
+
+local function registerFire(self)
     useImpact();
     if self.IsWrath() then
         self:RegisterAura("firestarter", 0, 54741, "impact", "Top", 0.8, 255, 255, 255, true, { (GetSpellInfo(2120)) }); -- May conflict with Impact location
@@ -587,10 +669,14 @@ local function registerClass(self)
         self:RegisterAura("hot_streak_full", 0, hotStreakSoDSpellID, "hot_streak", "Left + Right (Flipped)", 1, 255, 255, 255, true, { (GetSpellInfo(pyroblast)) });
     elseif self.IsCata() then
         self:RegisterAura("hot_streak_full", 0, hotStreakSpellID, "hot_streak", "Left + Right (Flipped)", 1, 255, 255, 255, true, { pyroblastBang });
+    elseif self.IsMoP() then
+        useHeatingUpAndHotStreak();
     else
         self:RegisterAura("hot_streak_full", 0, hotStreakSpellID, "hot_streak", "Left + Right (Flipped)", 1, 255, 255, 255, true, { (GetSpellInfo(pyroblast)) });
     end
-    self:RegisterAura("hot_streak_half", 0, heatingUpSpellID, "hot_streak", "Left + Right (Flipped)", 0.5, 255, 255, 255, false); -- Does not exist, but define it for option testing
+    if not self.IsMoP() then
+        self:RegisterAura("hot_streak_half", 0, heatingUpSpellID, "hot_streak", "Left + Right (Flipped)", 0.5, 255, 255, 255, false); -- Does not exist, but define it for option testing
+    end
     if not self.IsCata() then
         self:RegisterAura("hot_streak_duo", 0, hotStreakHeatingUpSpellID, "hot_streak", "Left + Right (Flipped)", 0.5, 255, 255, 255, false); -- Does not exist, but define it for option testing
         self:RegisterAura("hot_streak_duo", 0, hotStreakHeatingUpSpellID, "hot_streak", "Left + Right (Flipped)", 1, 255, 255, 255, true); -- Does not exist, but define it for option testing
@@ -598,8 +684,9 @@ local function registerClass(self)
     -- Heating Up (spellID == 48107) doesn't exist in Wrath Classic, so we can't use the above aura
     -- Instead, we track Fire Blast, Fireball, Living Bomb and Scorch non-periodic critical strikes
     -- Please look at HotStreakHandler and customCLEU for more information
+end
 
-    -- Frost Procs
+local function registerFrost(self)
     if self.IsSoD() then
         local iceLanceAndDeepFreezeSoD = { (GetSpellInfo(FrozenHandler.ice_lance_sod[1])), (GetSpellInfo(FrozenHandler.deep_freeze_sod[1])) };
         self:RegisterAura("fingers_of_frost_1_sod", 1, 400670, "frozen_fingers", "Left", 1, 255, 255, 255, true, iceLanceAndDeepFreezeSoD);
@@ -615,6 +702,8 @@ local function registerClass(self)
         self:RegisterAura("fingers_of_frost_1", 1, 44544, "frozen_fingers", "Left (CCW)", 1.1, 222, 222, 222, true, iceLanceAndDeepFreeze);
         self:RegisterAura("fingers_of_frost_2", 2, 44544, "frozen_fingers", "Left (CCW)", 1.1, 222, 222, 222, true, iceLanceAndDeepFreeze);
         self:RegisterAura("fingers_of_frost_2", 2, 44544, "frozen_fingers", "Right (CW)", 1.1, 222, 222, 222, true); -- no need to re-glow iceLanceAndDeepFreeze for right texture
+    elseif self.IsMoP() then
+        useFingersOfFrost();
     end
     if not self.IsCata() then
         self:RegisterAura("freeze", 0, FrozenHandler.fakeSpellID, FrozenHandler.saoTexture, "Top (CW)", FrozenHandler.saoScaleFactor, 255, 255, 255, false);
@@ -624,19 +713,21 @@ local function registerClass(self)
     if self.IsSoD() then
         self:RegisterAura("brain_freeze", 0, 400730, "brain_freeze", "Top", 1, 255, 255, 255, true, { (GetSpellInfo(133)), (GetSpellInfo(412532)), (GetSpellInfo(401502)) });
     elseif self.IsWrath() then
-        self:RegisterAura("brain_freeze", 0, 57761, "brain_freeze", "Top", 1, 255, 255, 255, true, { (GetSpellInfo(133)), (GetSpellInfo(44614)) });
+        self:RegisterAura("brain_freeze", 0, 57761, "brain_freeze", "Top", 1, 255, 255, 255, true, { (GetSpellInfo(133)), (GetSpellInfo(frostfireBolt)) });
     elseif self.IsCata() then
-        self:RegisterAura("brain_freeze", 0, 57761, "brain_freeze", "Top (CW)", 1, 255, 255, 255, true, { (GetSpellInfo(133)), (GetSpellInfo(44614)) });
+        self:RegisterAura("brain_freeze", 0, 57761, "brain_freeze", "Top (CW)", 1, 255, 255, 255, true, { (GetSpellInfo(133)), (GetSpellInfo(frostfireBolt)) });
+    elseif self.IsMoP() then
+        useBrainFreeze();
     end
+end
 
-    -- Arcane Procs
+local function registerArcane(self)
+    useArcaneMissiles();
     if self.IsSoD() then
     	-- Blue-ish, slightly smaller, to avoid confusion and overlap with Arcane Blast
         self:RegisterAura("missile_barrage", 0, 400589, "arcane_missiles", "Left + Right (Flipped)", 0.8, 103, 184, 238, true, { (GetSpellInfo(5143)) });
     elseif self.IsWrath() then
         self:RegisterAura("missile_barrage", 0, 44401, "arcane_missiles", "Left + Right (Flipped)", 1, 255, 255, 255, true, { (GetSpellInfo(5143)) });
-    elseif self.IsCata() then
-        useArcaneMissiles();
     end
     if self.IsCata() then
         local arcanePotency1 = 57529;
@@ -674,6 +765,12 @@ local function registerClass(self)
             self:RegisterAura("serendipity_sod", nbStacks, arcaneBlastSoDBuff, texture, "Left + Right (Flipped)", scale, 255, 255, 255, pulse, glowIDs);
         end
     end
+end
+
+local function registerClass(self)
+    registerArcane(self);
+    registerFire(self);
+    registerFrost(self);
 end
 
 local function loadOptions(self)
@@ -715,7 +812,6 @@ local function loadOptions(self)
     local arcaneExplosion = 1449;
     local flamestrike = 2120;
     local fireball = 133;
-    local frostfireBolt = 44614;
     local frostfireBoltSoD = 401502;
     local spellfrostBoltSoD = 412532;
     local iceLance = FrozenHandler.ice_lance[1];